Extract deepEqual to its own module
